Fixes multi group issues with job matching.
The issue was related to how these scripts expect the first group match
to be the group that the node boot as. Due to the Compute ndoes on
Chicoma being split into two groups (Compute_CPU and Compute_GPU) along
with other groups, nodes now have more than one group they are a member
of and thus this issue.